Fixed GCC warnings: -Wconversion -Wshadow -Wunused-parameter. JIRA:CXX-5413
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@74559 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objmgr/split/size.hpp
@@ -96,7 +96,7 @@ public:
         }
     double GetRatio(void) const
         {
-            return double(m_ZipSize)/m_AsnSize;
+            return (double)m_ZipSize/(double)m_AsnSize;
         }
 
     CNcbiOstream& Print(CNcbiOstream& out) const;
